Do not wait for messages to send before returning response
This deals with a recent issue at A&A whereby their gateway no longer
seems to be able to deal with messages being submitted at the same time
as they are being received. Before this commit submission of messages
was hanging because there was no response whilst the gateway was waiting
for a response from this script.